Create special_matrix
A square matrix is called special matrix if all the non-diagonal elements of the matrix are zero and the diagonal elements are non-zero.

You are given a square matrix with N rows and columns. You have to check whether the given matrix is special or not.

NOTE: You need to complete the given function. The input and printing of output will be handled by the driver code.

Input Format:
The first line contains the number of test cases.

For each test case: The first line contains N.

The next N lines contain N integers each representing the elements of the matrix.

Output Format:
For each test case in t, output t lines true if the matrix is special else false

Example 1:
Input:

1
3
1 0 2
0 2 0
3 0 1
Output:

true
Explanation:

The diagonal elements are non-zero and non-diagonal elements are zero.
Example 2:
Input:

1
3
1 0 1
1 2 0
2 0 3
Output:

false
Explanation:

The non-diagonal element in second row of first column is non-zero.

Constraints:
1 <= T <= 10

1 <= N <= 200

0 <= A[i] <=1000000